feat(insights): add insights feature flag
