+++ packages/server/graphql/public/typeDefs/_legacy.graphql
@@ -3215,26 +3215,6 @@ enum TaskStatusEnum {
 }
 
 """
-The types of flags that give an individual user super powers
-"""
-type UserFeatureFlags {
-  """
-  true if spotlight is allowed
-  """
-  spotlight: Boolean!
-
-  """
-  true if standups is allowed
-  """
-  standups: Boolean!
-
-  """
-  true if gitlab is allowed
-  """
-  gitlab: Boolean!
-}
-
-"""
 A monthly billing invoice for an organization
 """
 type Invoice {
@@ -7099,26 +7079,6 @@ type Mutation {
   addReflectTemplatePrompt(templateId: ID!): AddReflectTemplatePromptPayload
   addSlackAuth(code: ID!, teamId: ID!): AddSlackAuthPayload!
 
-  """
-  Give someone advanced features in a flag
-  """
-  addFeatureFlag(
-    """
-    a list of the complete or partial email of the person to whom you are giving advanced features.
-          Matches via a regex to support entire domains
-    """
-    emails: [String!]
-
-    """
-    grant access to an entire domain. the part of the email after the @
-    """
-    domain: String
-
-    """
-    the flag that you want to give to the user
-    """
-    flag: UserFlagEnum!
-  ): AddFeatureFlagPayload!
   addGitHubAuth(code: ID!, teamId: ID!): AddGitHubAuthPayload!
 
   """
@@ -8691,29 +8651,6 @@ type AddSlackAuthPayload {
   The user with updated slackAuth
   """
   user: User
-}
-
-type AddFeatureFlagPayload {
-  error: StandardMutationError
-
-  """
-  the user that was given the super power. Use users instead in GraphiQL since it may affect multiple users
-  """
-  user: User
-
-  """
-  the users given the super power
-  """
-  users: [User]
-}
-
-"""
-A flag to give an individual user super powers
-"""
-enum UserFlagEnum {
-  spotlight
-  standups
-  gitlab
 }
 
 type AddGitHubAuthPayload {
@@ -11128,7 +11065,6 @@ type UpdateDragLocationPayload {
 
 union NotificationSubscriptionPayload =
     AcceptTeamInvitationPayload
-  | AddFeatureFlagPayload
   | AddNewFeaturePayload
   | AddOrgPayload
   | AddTeamPayload
